[code] Split header and implementation
