+++ include/inexor/vulkan-renderer/gltf-model/mesh.hpp
@@ -31,18 +31,12 @@ struct Mesh {
 
     /// @brief Sets the model matrix.
     /// @param mat [in] The input matrix.
-    void set_matrix(const glm::mat4 &mat) {
-        uniform_block.matrix = mat;
-    }
+    void set_matrix(const glm::mat4 &mat);
 
     /// @brief Sets the bounding box of the model.
     /// @param min [in] The minimum vector (edge of the bounding box)
     /// @param max [in] The maximum vector (edge of the bounding box)
-    void set_bounding_box(glm::vec3 min, glm::vec3 max) {
-        bb.min = min;
-        bb.max = max;
-        bb.valid = true;
-    }
+    void set_bounding_box(glm::vec3 min, glm::vec3 max);
 };
 
 } // namespace inexor::vulkan_renderer::gltf_model
